# app.py
from flask import Flask, jsonify, request, send_from_directory, send_file, Response
from lib import CollectionRegistry, TerminologyRegistry, MappingRegistry, \
    ApiError, NotFound, ValidationError, createTripleStore
from werkzeug.exceptions import HTTPException
import os
import json
from datetime import datetime, timezone
from pathlib import Path
from flask_cors import CORS


app = Flask(__name__)
CORS(app)
app.json.compact = False

collections = None
terminologies = None
mappings = None


def configure(**config):
    global collections
    global terminologies
    global mappings

    title = config.get('title', os.getenv('TITLE', 'Grimpo Knowledge Graph Importer'))

    app.config['title'] = title
    app.config['base'] = config.get('base', os.getenv('BASE', 'http://example.org/'))
    app.config['frontend'] = config.get('frontend', os.getenv(
        'FRONTEND', app.config['base']))
    app.config['sparql'] = config.get('sparql', os.getenv('SPARQL'))
    app.config['stage'] = config.get('stage', os.getenv('STAGE', 'stage'))
    app.config['data'] = config.get('data', os.getenv('DATA', 'data'))
    app.config['store'] = createTripleStore(app.config.get("sparql"))

    terminologies = TerminologyRegistry(**app.config)
    collections = CollectionRegistry(**app.config, terminologies=terminologies)
    mappings = MappingRegistry(**app.config)


with app.app_context():
    configure()


@app.errorhandler(ApiError)
def handle_apierror(e):
    return jsonify(e.to_dict()), type(e).code


@app.errorhandler(ValidationError)
def handle_validationerror(e):
    e = e.to_dict()
    e["code"] = 400
    return jsonify(e), 400


@app.errorhandler(HTTPException)
def handle_exception(e):
    return jsonify({
        "code": e.code,
        "name": e.name,
        "message": e.description,
    }), e.code


def route(method, path, fn):
    fn.__name__ = f'{method}-{path}'
    app.add_url_rule(path, methods=[method], view_func=fn)


def api(method, path, fn):
    route(method, path, lambda *args, **kws: jsonify(fn(*args, **kws)))


@app.route('/status.json')
def status():
    # TODO: Fully adjust to OpenAPI specification and move statistics to summary
    values = {key: val for key, val in app.config.items() if key.islower() and type(val) in [str, bool]}
    try:
        values["collections"] = len(collections.registered())
        values["terminologies"] = len(terminologies.registered())
        values["mappings"] = len(mappings.registered())
        values['connected'] = True
    except Exception as e:
        values['connected'] = False
    return values

@app.route('/openapi.json')
def openapi():
    openapi = json.load(open("openapi.json", "r"))
    for kind in ["collection", "terminology", "mappings"]:
        with open(f"lib/{kind}-schema.json", "r") as file:
            openapi["components"]["schemas"][kind] = json.load(file)
    return openapi
 
@app.route('/metadata')
def metadata():
    metadata = {
        "collections": collections.list(),
        "mappings": mappings.list(),
        "terminologies": [{"uri": t["uri"]} for t in terminologies.list()]
    }
    return metadata


route('GET', '/', lambda: send_file("ui/index.html"))

for file in Path('ui').glob('*.*'):
    route('GET', f'/{file.name}', lambda f=file: send_file(str(f)))

if not app.config.get('sparql'):
    route('GET', '/sparql', lambda: app.config['store'].query_request(request))
    route('POST', '/sparql', lambda: app.config['store'].query_request(request))

api('GET', '/data/', lambda: list_files(Path(app.config['data'])))

api('GET', '/terminologies/', lambda: terminologies.list())
api('GET', '/terminologies/namespaces.json', lambda: terminologies.namespaces())

route('GET', '/terminologies/skosmos.ttl', lambda: Response(terminologies.skosmos(), mimetype="text/turtle"))

api('PUT', '/terminologies/', lambda: terminologies.replace(request.get_json(force=True)))
api('GET', '/terminologies/<int:id>', lambda id: terminologies.get(id))
api('PUT', '/terminologies/<int:id>', lambda id: terminologies.register({"id": str(id)}))
api('DELETE', '/terminologies/<int:id>', lambda id: terminologies.delete(id))
api('POST', '/terminologies/<int:id>/receive', lambda id: terminologies.receive(id, request.args.get('from', None)))
api('GET', '/terminologies/<int:id>/receive', lambda id: terminologies.receive_log(id))
api('GET', '/terminologies/<int:id>/load', lambda id: terminologies.load_log(id))
api('POST', '/terminologies/<int:id>/load', lambda id: terminologies.load(id))
api('POST', '/terminologies/<int:id>/remove', lambda id: terminologies.remove(id))

api('GET', '/collections/', lambda: collections.list())
api('GET', '/collections/schema.json', lambda: collections.schema)
api('PUT', '/collections/', lambda: collections.replace(request.get_json(force=True)))
api('POST', '/collections/', lambda: collections.register(request.get_json(force=True)))
api('GET', '/collections/<int:id>', lambda id: collections.get(id))
api('PUT', '/collections/<int:id>', lambda id: collections.register(request.get_json(force=True), id))
api('DELETE', '/collections/<int:id>', lambda id: collections.delete(id))
api('POST', '/collections/<int:id>/receive', lambda id: collections.receive(id, request.args.get("from", None)))
api('GET', '/collections/<int:id>/receive', lambda id: collections.receive_log(id))
api('POST', '/collections/<int:id>/load', lambda id: collections.load(id))
api('POST', '/collections/<int:id>/add', lambda id: collections.load(id, add=True))
api('GET', '/collections/<int:id>/load', lambda id: collections.load_log(id))
api('POST', '/collections/<int:id>/remove', lambda id: collections.remove(id))

api('GET', '/mappings/', lambda: mappings.list())
api('GET', '/mappings/schema.json', lambda: mappings.schema)
api('GET', '/mappings/properties.json', lambda: mappings.properties)
api('PUT', '/mappings/', lambda: mappings.replace(request.get_json(force=True)))
api('POST', '/mappings/', lambda: mappings.register(request.get_json(force=True)))
api('GET', '/mappings/<int:id>', lambda id: mappings.get(id))
api('PUT', '/mappings/<int:id>', lambda id: mappings.register(request.get_json(force=True), id))
api('DELETE', '/mappings/<int:id>', lambda id: mappings.delete(id))
api('POST', '/mappings/<int:id>/append', lambda id: mappings.append(id, request.get_data()))
api('POST', '/mappings/<int:id>/detach', lambda id: mappings.detach(id, request.get_data()))
api('POST', '/mappings/<int:id>/receive', lambda id: mappings.receive(id, request.args.get("from", None)))
api('GET', '/mappings/<int:id>/receive', lambda id: mappings.receive_log(id))
api('POST', '/mappings/<int:id>/load', lambda id: mappings.load(id))
api('GET', '/mappings/<int:id>/load', lambda id: mappings.load_log(id))
api('POST', '/mappings/<int:id>/remove', lambda id: mappings.remove(id))


def list_files(dir):
    files = []
    if dir.is_dir():
        for file in [f for f in dir.iterdir() if f.is_file()]:
            st = file.stat()
            files.append({
                "name": file.name,
                "size": st.st_size,
                "created": datetime.fromtimestamp(st.st_ctime, tz=timezone.utc).isoformat(),
                "modified": datetime.fromtimestamp(st.st_mtime, tz=timezone.utc).isoformat(),
            })
    return sorted(files, key=lambda f: f["modified"])


def serve_dir(dir, root, filename=None, id=None):
    if filename:
        file = dir / filename
        if "/" in filename or not file.is_file():
            raise NotFound("File not found!")
        return send_from_directory(dir, filename)
    else:
        return list_files(dir)


def stage(kind, id, filename=None):
    dir = Path(app.config["stage"]) / kind / str(id)
    if dir.is_dir():
        return serve_dir(dir, "../../../", filename, id)
    else:
        raise NotFound(f"{kind} {id} not found!")


@app.route('/terminologies/<int:id>/stage/')
@app.route('/terminologies/<int:id>/stage/<filename>')
def terminology_stage(id, filename=None):
    return stage("terminologies", id, filename)


@app.route('/collections/<int:id>/stage/')
@app.route('/collections/<int:id>/stage/<filename>')
def collection_stage(id, filename=None):
    return stage("collections", id, filename)


@app.route('/mappings/<int:id>/stage/')
@app.route('/mappings/<int:id>/stage/<filename>')
def mappings_stage(id, filename=None):
    return stage("mappings", id, filename)


@app.route('/data/<filename>')
def data_directory(filename=None):
    return serve_dir(Path(app.config["data"]), "../", filename)