# openapi.json
{
  "openapi": "3.2.1",
  "info": {
    "title": "Grimpo Knowledge Graph Importer API",
    "version": "1.0.0"
  },
  "components": {
...
    "parameters": {
      "from": {
        "name": "from",
        "description": "URL or file",
        "in": "query",
        "schema": {
          "type": "string"
        }
      }
    },
    "requestBodies": {
      "collection": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "'#/components/schemas/collection"
            }
          }
        }
      },
      "collections": {
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "'#/components/schemas/collection"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/terminologies/{id}/receive": {
      "post": {
        "summary": "Receive terminology data",
        "parameters": [
          {
            "$ref": "'#/components/parameters/from"
          }
        ]
      },
      "get": {
        "summary": "Get latest receive report of a terminology"
      }
    },
...
    "/collections": {
      "get": {
        "summary": "Return the list of registered collections"
      },
      "put": {
        "summary": "Register a list of collections, replacing existing ones",
        "requestBody": {
          "$ref": "'#/components/requestBodies/collections"
        }
      },
      "post": {
        "summary": "Register a new collection or update metadata of a registered collection",
        "requestBody": {
          "$ref": "'#/components/requestBodies/collection"
        }
      }
    },
...
    "/collections/{id}": {
      "get": {
        "summary": "Return metadata of a specific registered collection"
      },
      "put": {
        "summary": "Update metadata of a specific registered collection or register a new collection",
        "requestBody": {
          "$ref": "'#/components/requestBodies/collection"
        }
      },
      "delete": {
        "summary": "Unregister a collection and remove it from the triple store and staging area"
      }
    },
...
    "/collections/{id}/receive": {
      "post": {
        "summary": "Receive and process collection data",
        "parameters": [
          {
            "$ref": "'#/components/parameters/from"
          }
        ]
      },
      "get": {
        "summary": "Get latest receive log of a collection"
      }
    },
...
  }
}